pyproject.toml
[tool.poetry]
name = "django-payments-flow"
version = "0.1.3"
description = "Soporte Flow para Django Payments"
authors = ["Mario Hernandez <mariofix@proton.me>"]
license = "MIT"
readme = "README.md"
packages = [{include = "django_payments_flow"}]
repository = "https://github.com/mariofix/django-payments-flow"
documentation = "https://mariofix.github.io/django-payments-flow/"
homepage = "https://www.flow.cl/docs/api.html"
keywords = [
    "flow",
    "pagos",
    "django",
    "payment",
    "django-payments"
]
classifiers=[
    "Development Status :: 2 - Pre-Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: Django",
    "Operating System :: OS Independent"
]

[tool.poetry.dependencies]
python = "^3.8.10"
django-payments = "2.0.0"
pyflowcl = "1.1.2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
black = "^23.3.0"
isort = "^5.12.0"
coverage = "^7.2.7"
pytest = "^7.4.0"
faker = ">=18.13,<20.0"
mkdocs = "^1.4.3"
mkdocs-material = "^9.1.18"
mkdocstrings = {extras = ["python"], version = "^0.22.0"}
mkdocs-git-revision-date-localized-plugin = "^1.2.0"
mkdocs-git-committers-plugin-2 = "^1.1.2"
pillow = "^10.0.0"
cairosvg = "^2.7.0"
lxml = "^4.9.3"
pytest-cov = "^4.1.0"
pytest-django = "^4.5.2"
pytest-mock = "^3.11.1"
dj-database-url = "^2.0.0"
factory-boy = "^3.3.0"
tox = "^4.6.4"
tox-gh-actions = "^3.1.3"



[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.django_settings"
addopts = "--cov=django_payments_flow/ --cov-branch --cov-report=term-missing --cov-report=xml"

[tool.black]
line-length = 119
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 119
multi_line_output = 5
py_version = 38
src_paths = ["django_payments_flow", "tests", "sandbox"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
